requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
executing==2.2.0
filelock==3.13.1
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.7.0
gdown==5.2.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
huggingface-hub==0.33.2
icecream==2.1.5
idna==3.10
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.10.3
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mpmath @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mpmath_1699484863771/work
multidict==6.6.4
munch==4.0.0
networkx==3.3
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.0
pillow==11.0.0
propcache==0.3.2
psutil==7.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
requests @ file:///C:/b/abs_f53smfvswb/croot/requests_1756709376206/work
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
segmentation_models_pytorch==0.5.0
setuptools==72.1.0
simsimd==6.5.1
six==1.17.0
sklearn_quantile==0.1.1
soupsieve==2.7
stringzilla==3.12.6
sympy==1.13.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.19
torch==2.5.1
torch-geometric==2.6.1
torch_cluster==1.6.3+pt25cu118
torch_scatter==2.1.2+pt25cu118
torch_sparse==0.6.18+pt25cu118
torch_spline_conv==1.2.2+pt25cu118
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wheel==0.45.1
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
yarl==1.20.1
